requirements.txt
aiofiles==23.2.1
aiosqlite==0.19.0
anyio==3.7.1
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.104.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
iso8601==2.1.0
Jinja2==3.1.0
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.5.0
pydantic==1.10.0
PyJWT==2.8.0
PyPika==0.48.9
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tortoise-orm==0.16.6
typing_extensions==4.8.0
uvicorn==0.15.0
